move apps_utilities.h to utilities.h
