+++ apps/utilities.h
@@ -15,11 +15,21 @@
 #include <deal.II/base/utilities.h>
 
 #include <deal.II/lac/full_matrix.h>
+#include <deal.II/lac/la_parallel_vector.h>
+
 
 #include "git_version.h"
 #include "solvers_and_preconditioners/TPSS/generic_functionalities.h"
+#include "solvers_and_preconditioners/TPSS/tensors.h"
 
+
+#include <algorithm>
+#include <cstring>
 #include <iostream>
+#include <string>
+#include <vector>
+
+
 
 using namespace dealii;
 
@@ -67,6 +77,58 @@ generic_info_to_fstring()
                                     8 * size_of_global_dof_index);
   return oss.str();
 }
+
+
+
+static constexpr char const * skipper = "o";
+
+std::vector<char const *>
+args_to_strings(const int argc_in, char * argv_in[])
+{
+  std::vector<char const *> tmp;
+  std::copy_n(argv_in, argc_in, std::back_inserter(tmp));
+  return tmp;
+}
+
+struct ConditionalAtoi
+{
+  ConditionalAtoi(const int argc_in, char * argv_in[]) : argv(args_to_strings(argc_in, argv_in))
+  {
+  }
+
+  template<typename T>
+  void
+  operator()(T & prm, const std::size_t index)
+  {
+    if(argv.size() <= index)
+      return;
+    if(std::strcmp(argv[index], skipper) == 0)
+      return;
+    prm = std::atoi(argv[index]);
+  }
+
+  std::vector<char const *> argv;
+};
+
+struct ConditionalAtof
+{
+  ConditionalAtof(const int argc_in, char * argv_in[]) : argv(args_to_strings(argc_in, argv_in))
+  {
+  }
+
+  template<typename T>
+  void
+  operator()(T & prm, const std::size_t index)
+  {
+    if(argv.size() <= index)
+      return;
+    if(std::strcmp(argv[index], skipper) == 0)
+      return;
+    prm = std::atof(argv[index]);
+  }
+
+  std::vector<char const *> argv;
+};
